removed unnecssary/problematic _i2cPort->end();
This was originally thought to help inconsistency issues when using the setI2CAddress() function, but it is actually not necessary. It was also causing compilation error when using the ESP8266 or ESP32. This has been verified to compile and work on both hardware.

Also, added a comment to explain why there is a delay in there.